+++ packages/meedu/CHANGELOG.md
@@ -1,3 +1,5 @@
+## [4.2.0]
+- Added autoRemove and onRemove in dependency injection.
 ## [4.1.0]
 - Renamed `onDispose` to `dispose`.
 - Added `onlyUpdate` method in StateNotifier.